Limpiando y agregando loader, tambien corriegien las imagenes en el chat
