+++ Frontend/src/Components/Loader/LoaderFull.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import {Loader, Cohete} from './StyledLoader'
+
+export default function LoaderFull( ) {
+
+
+    return (
+        
+    <div>
+        <Loader></Loader>
+        <Cohete></Cohete>
+    </div>
+  
+    )};
+
+  
+  
+
