+++ Frontend/src/Components/PostContainer/PostContainer.jsx
@@ -1,4 +1,4 @@
-import { Loader } from '../Loader/Loader';
+import  LoaderFull  from '../Loader/LoaderFull';
 import Post from '../Post/Post'
 import { StyledPostContainer } from "./styles";
 
@@ -7,7 +7,7 @@ export default function PostContainer({ posts = [] }) {
     return (
         <StyledPostContainer>
             {posts.length ? posts.map((p) => <Post p={p} key={p._id} />) :
-                <Loader></Loader>}
+                <LoaderFull></LoaderFull>}
         </ StyledPostContainer>
     )
 }
