+++ Frontend/src/Components/Profile/Profile.jsx
@@ -12,7 +12,7 @@ import {
 import CardProfile from "./CardProfile";
 import Layout from "../Layout/Layout";
 import PostContainer from "../PostContainer/PostContainer";
-import { Loader } from "../Loader/Loader";
+import LoaderFull from "../Loader/LoaderFull";
 
 const Profile = () => {
   const { id } = useParams();
@@ -38,7 +38,7 @@ const Profile = () => {
       {Object.keys(profile).length && Object.keys(myId).length ? (
         <CardProfile profile={profile} followUser={followUser} myId={myId} />
       ) : (
-        <Loader></Loader>
+        <LoaderFull></LoaderFull>
       )}
       <PostContainer posts={myFriendPost} />
     </Layout>
